Decimal Hidden Values Fixed
Datatype Double has 15 digits of precision after the decimal but its displayed as 4 digits after decimal. For example 11.189656743267890 will shown as 11.1896. This caused a lot of correct answers to be shown as wrong. For example 2.4509 * 4.56655 AND 5.26 *  2.1273 will both be printed as 11.1896 but actually the decimals diverge after 4 places and the program will show them both to be unequal despite being printed as equal.
This update fixes this issue, Now both the values will be measured upto 3 digits after decimal only and others will be discarded. Hence 11.1896 answer of 1st product and 11.1896 answer of 2nd product will be equal in value.